+++ h5netcdf/core.py
@@ -64,6 +64,43 @@ def _invalid_netcdf_feature(feature, allow):
         raise CompatibilityError(msg)
 
 
+def _expanded_indexer(key, ndim):
+    """Expand indexing key to tuple of slices with length equal the number of dimensions."""
+    # always return tuple and force colons to slices
+    key = np.index_exp[key]
+
+    # dimensions
+    len_key = len(key)
+
+    # find Ellipsis
+    ellipsis = [i for i, k in enumerate(key) if k is Ellipsis]
+    if len(ellipsis) > 1:
+        raise IndexError(
+            f"an index can only have a single ellipsis ('...'), {len(ellipsis)} given"
+        )
+    else:
+        # expand Ellipsis wherever it is
+        len_key -= len(ellipsis)
+        res_dim_cnt = ndim - len_key
+        res_dims = res_dim_cnt * (slice(None),)
+        ellipsis = ellipsis[0] if ellipsis else None
+
+    # check for correct dimensionality
+    if ndim and res_dim_cnt < 0:
+        raise IndexError(
+            f"too many indices for array: array is {ndim}-dimensional, but {len_key} were indexed"
+        )
+
+    # convert integer indices to slices
+    key = tuple([slice(k, k + 1) if isinstance(k, int) else k for k in key])
+
+    # slices to build resulting key
+    k1 = slice(ellipsis)
+    k2 = slice(len_key, None) if ellipsis is None else slice(ellipsis + 1, None)
+
+    return key[k1] + res_dims + key[k2]
+
+
 class BaseVariable(object):
     def __init__(self, parent, name, dimensions=None):
         self._parent_ref = weakref.ref(parent)
@@ -108,7 +145,7 @@ class BaseVariable(object):
         phony_dims = defaultdict(int)
         for axis, dim in enumerate(self._h5ds.dims):
             # get current dimension
-            dimsize = self.shape[axis]
+            dimsize = self._h5ds.shape[axis]
             phony_dims[dimsize] += 1
             if len(dim):
                 name = _name_from_dimension(dim)
@@ -148,6 +185,42 @@ class BaseVariable(object):
             if "_Netcdf4Dimid" in dim.attrs:
                 self._h5ds.attrs["_Netcdf4Dimid"] = dim.attrs["_Netcdf4Dimid"]
 
+    def _maybe_resize_dimensions(self, key, value):
+        """Resize according to given key with respect to variable dimensions"""
+        # expand key to slices
+        key = _expanded_indexer(key, self.ndim)
+        new_shape = ()
+        v = None
+        for i, dim in enumerate(self.dimensions):
+            # is unlimited dimensions
+            if not self._parent._dim_sizes[dim]:
+                if key[i].stop is None:
+                    # if stop is None, get dimensions from value,
+                    # they must match with variable dimension
+                    if v is None:
+                        v = np.asarray(value)
+                    if v.ndim == self.ndim:
+                        new_max = max(v.shape[i], self._h5ds.shape[i])
+                    elif v.ndim == 0:
+                        # for scalars we take the current dimension size
+                        new_max = self._parent._current_dim_sizes[dim]
+                    else:
+                        raise IndexError("shape of data does not conform to slice")
+                else:
+                    new_max = max(key[i].stop, self._h5ds.shape[i])
+                # resize unlimited dimension if needed but no other variables
+                # this is in line with `netcdf4-python` which only resizes
+                # the dimension and this variable
+                if self._parent._current_dim_sizes[dim] < new_max:
+                    self._parent.resize_dimension(dim, new_max, resize_vars=False)
+                new_shape += (new_max,)
+            else:
+                new_shape += (self._parent._current_dim_sizes[dim],)
+
+        # increase variable size if shape is changing
+        if self._h5ds.shape != new_shape:
+            self._h5ds.resize(new_shape)
+
     @property
     def dimensions(self):
         if self._dimensions is None:
@@ -156,7 +229,8 @@ class BaseVariable(object):
 
     @property
     def shape(self):
-        return self._h5ds.shape
+        # return actual dimensions sizes, this is in line with netcdf4-python
+        return tuple([self._parent._current_dim_sizes[d] for d in self.dimensions])
 
     @property
     def ndim(self):
@@ -177,9 +251,28 @@ class BaseVariable(object):
             string_info = h5py.check_string_dtype(self._h5ds.dtype)
             if string_info and string_info.length is None:
                 return self._h5ds.asstr()[key]
+
+        # return array padded with fillvalue
+        if self.dtype != str and self.dtype.kind in ["f", "i", "u"]:
+            sdiff = [d0 - d1 for d0, d1 in zip(self.shape, self._h5ds.shape)]
+            if sum(sdiff):
+                fv = self.dtype.type(self._h5ds.fillvalue)
+                padding = [(0, s) for s in sdiff]
+                return np.pad(
+                    self._h5ds,
+                    pad_width=padding,
+                    mode="constant",
+                    constant_values=fv,
+                )[key]
+
         return self._h5ds[key]
 
     def __setitem__(self, key, value):
+        from .legacyapi import Dataset
+
+        # resize on write only for legacy API
+        if isinstance(self._parent._root, Dataset):
+            self._maybe_resize_dimensions(key, value)
         self._h5ds[key] = value
 
     @property
@@ -751,46 +844,43 @@ class Group(Mapping):
         header = "<%s %r (%s members)>" % (self._cls_name, self.name, len(self))
         return "\n".join([header] + self._repr_body())
 
-    def resize_dimension(self, dimension, size, resize_vars=True):
-        """
-        Resize a dimension to a certain size.
+    def resize_dimension(self, dim, size, resize_vars=True):
+        """Resize a dimension to a certain size.
 
         It will pad with the underlying HDF5 data sets' fill values (usually
         zero) where necessary.
         """
-        if self.dimensions[dimension] is not None:
+        if self.dimensions[dim] is not None:
             raise ValueError(
-                "Dimension '%s' is not unlimited and thus "
-                "cannot be resized." % dimension
+                "Dimension '%s' is not unlimited and thus cannot be resized." % dim
             )
 
         # Resize the dimension.
-        self._current_dim_sizes[dimension] = size
+        self._current_dim_sizes[dim] = size
 
-        # Resize dimension in file
-        if dimension in self._h5group:
-            self._h5group[dimension].resize((size,))
+        # Needs resize in any case, if variable, if dimensions and if coordinate
+        if dim in self._h5group:
+            self._h5group[dim].resize((size,))
 
         # Adapt recursively to all variables
         if resize_vars:
-            self._resize_variables(dimension, size)
+            self._resize_variables(dim, size)
 
-    def _resize_variables(self, dimension, size, recurse=True):
+    def _resize_variables(self, dim, size, recurse=True):
         """Resize variables"""
-        # Adapt to all variables
         for var in self.variables.values():
-            new_shape = list(var.shape)
+            new_shape = list(var._h5ds.shape)
             for i, d in enumerate(var.dimensions):
-                if d == dimension:
+                if d == dim:
                     new_shape[i] = size
             new_shape = tuple(new_shape)
-            if new_shape != var.shape:
+            if new_shape != var._h5ds.shape:
                 var._h5ds.resize(new_shape)
 
         # Recurse as dimensions are visible to this group and all child groups.
         if recurse:
             for i in self.groups.values():
-                i._resize_variables(dimension, size)
+                i._resize_variables(dim, size)
 
 
 class File(Group):