+++ h5netcdf/tests/test_h5netcdf.py
@@ -68,6 +68,11 @@ def decode_vlen_strings(request):
         return {}
 
 
+@pytest.fixture(params=[netCDF4, legacyapi])
+def netcdf_write_module(request):
+    return request.param
+
+
 def get_hdf5_module(resource):
     """Return the correct h5py module based on the input resource."""
     if isinstance(resource, str) and resource.startswith(remote_h5):
@@ -937,11 +942,15 @@ def test_creating_variables_with_unlimited_dimensions(tmp_local_or_remote_netcdf
             )
         assert e.value.args[0] == "Shape tuple is incompatible with data"
 
+        # Creating a coordinate variable
+        f.create_variable("x", dimensions=("x",), dtype=np.int64)
+
         # Resize data.
         assert f.variables["dummy"].shape == (0, 2)
         f.resize_dimension("x", 3)
         # This will also force a resize of the existing variables and it will
         # be padded with zeros..
+        assert f._current_dim_sizes["x"] == 3
         np.testing.assert_allclose(f.variables["dummy"], np.zeros((3, 2)))
 
         # Creating another variable with no data will now also take the shape
@@ -949,6 +958,17 @@ def test_creating_variables_with_unlimited_dimensions(tmp_local_or_remote_netcdf
         f.create_variable("dummy3", dimensions=("x", "y"), dtype=np.int64)
         assert f.variables["dummy3"].shape == (3, 2)
         assert f.variables["dummy3"]._h5ds.maxshape == (None, 2)
+        np.testing.assert_allclose(f.variables["dummy3"], np.zeros((3, 2)))
+
+        # Writing to a variable with an unlimited dimension raises
+        with pytest.raises(TypeError) as e:
+            f.variables["dummy3"][:] = np.ones((5, 2))
+        assert e.value.args[0] == "Can't broadcast (5, 2) -> (3, 2)"
+        assert f.variables["dummy3"].shape == (3, 2)
+        assert f.variables["dummy3"]._h5ds.maxshape == (None, 2)
+        assert f["x"].shape == (3,)
+        assert f._current_dim_sizes["x"] == 3
+        np.testing.assert_allclose(f.variables["dummy3"], np.zeros((3, 2)))
 
     # Close and read again to also test correct parsing of unlimited
     # dimensions.
@@ -967,6 +987,7 @@ def test_writing_to_an_unlimited_dimension(tmp_local_or_remote_netcdf):
         # Two dimensions, only one is unlimited.
         f.dimensions["x"] = None
         f.dimensions["y"] = 3
+        f.dimensions["z"] = None
 
         # Cannot create it without first resizing it.
         with pytest.raises(ValueError) as e:
@@ -979,6 +1000,7 @@ def test_writing_to_an_unlimited_dimension(tmp_local_or_remote_netcdf):
         f.create_variable("dummy1", dimensions=("x", "y"), dtype=np.int64)
         f.create_variable("dummy2", dimensions=("x", "y"), dtype=np.int64)
         f.create_variable("dummy3", dimensions=("x", "y"), dtype=np.int64)
+        f.create_variable("dummyX", dimensions=("x", "y", "z"), dtype=np.int64)
         g = f.create_group("test")
         g.create_variable("dummy4", dimensions=("y", "x", "x"), dtype=np.int64)
         g.create_variable("dummy5", dimensions=("y", "y"), dtype=np.int64)
@@ -986,20 +1008,22 @@ def test_writing_to_an_unlimited_dimension(tmp_local_or_remote_netcdf):
         assert f.variables["dummy1"].shape == (0, 3)
         assert f.variables["dummy2"].shape == (0, 3)
         assert f.variables["dummy3"].shape == (0, 3)
+        assert f.variables["dummyX"].shape == (0, 3, 0)
         assert g.variables["dummy4"].shape == (3, 0, 0)
         assert g.variables["dummy5"].shape == (3, 3)
+
+        # resize dimensions and all connected variables
         f.resize_dimension("x", 2)
         assert f.variables["dummy1"].shape == (2, 3)
         assert f.variables["dummy2"].shape == (2, 3)
         assert f.variables["dummy3"].shape == (2, 3)
+        assert f.variables["dummyX"].shape == (2, 3, 0)
         assert g.variables["dummy4"].shape == (3, 2, 2)
         assert g.variables["dummy5"].shape == (3, 3)
 
-        f.variables["dummy2"][:] = [[1, 2, 3], [5, 6, 7]]
-        np.testing.assert_allclose(f.variables["dummy2"], [[1, 2, 3], [5, 6, 7]])
-
-        f.variables["dummy3"][...] = [[1, 2, 3], [5, 6, 7]]
-        np.testing.assert_allclose(f.variables["dummy3"], [[1, 2, 3], [5, 6, 7]])
+        # broadcast writing
+        f.variables["dummy3"][...] = [[1, 2, 3]]
+        np.testing.assert_allclose(f.variables["dummy3"], [[1, 2, 3], [1, 2, 3]])
 
 
 def test_c_api_can_read_unlimited_dimensions(tmp_local_netcdf):
@@ -1059,7 +1083,9 @@ def test_reading_unlimited_dimensions_created_with_c_api(tmp_local_netcdf):
         assert f["dummy1"].shape == (2, 3)
         # XXX: This array has some data with dimension x - netcdf does not
         # appear to keep dimensions consistent.
-        assert f["dummy2"].shape == (3, 0, 0)
+        # With https://github.com/h5netcdf/h5netcdf/pull/103 h5netcdf will
+        # return a padded array
+        assert f["dummy2"].shape == (3, 2, 2)
         f.groups["test"]["dummy3"].shape == (3, 3)
         f.groups["test"]["dummy4"].shape == (0, 0)
 
@@ -1238,8 +1264,6 @@ def create_netcdf_dimensions(ds, idx):
     ship = g.createVariable("ship", "S1", ("ship", "ship_strlen"))
     sample = g.createVariable("sample", "i8", ("time", "sample"))
 
-    if isinstance(ds, legacyapi.Dataset):
-        g.resize_dimension("time", 10 + idx)
     time[:] = np.arange(10 + idx)
     data[:] = np.ones((3 + idx, 2 + idx, 10 + idx, 5 + idx)) * 12.0
     collide[...] = np.arange(5 + idx)
@@ -1338,10 +1362,7 @@ def check_netcdf_dimensions(tmp_netcdf, write_module, read_module):
             assert g.variables["data"].shape == (3 + i, 2 + i, 10 + i, 5 + i)
             assert g.variables["collide"].shape == (5 + i,)
             assert g.variables["non_collide"].shape == (5 + i,)
-            if write_module == netCDF4 and read_module in [legacyapi, h5netcdf]:
-                assert g.variables["sample"].shape == (2 + i, 2 + i)
-            else:
-                assert g.variables["sample"].shape == (10 + i, 2 + i)
+            assert g.variables["sample"].shape == (10 + i, 2 + i)
             assert g.variables["ship"].shape == (3 + i, 10)
 
 
@@ -1391,6 +1412,77 @@ def test_no_circular_references(tmp_local_netcdf):
     gc.collect()
     with h5netcdf.File(tmp_local_netcdf, "r") as ds:
         assert len(gc.get_referrers(ds)) == 1
+
+
+def test_expanded_variables_netcdf4(tmp_local_netcdf, netcdf_write_module):
+    with netcdf_write_module.Dataset(tmp_local_netcdf, "w") as ds:
+        f = ds.createGroup("test")
+        f.createDimension("x", None)
+        f.createDimension("y", 3)
+
+        dummy1 = f.createVariable("dummy1", float, ("x", "y"))
+        dummy2 = f.createVariable("dummy2", float, ("x", "y"))
+        dummy3 = f.createVariable("dummy3", float, ("x", "y"))
+        dummy4 = f.createVariable("dummy4", float, ("x", "y"))
+
+        dummy1[:] = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
+        dummy2[:] = [[1, 2, 3]]
+        dummy3[:] = [[1, 2, 3], [4, 5, 6]]
+
+        # don't mask, since h5netcdf doesn't do masking
+        if netcdf_write_module == netCDF4:
+            ds.set_auto_mask(False)
+
+        res1 = dummy1[:]
+        res2 = dummy2[:]
+        res3 = dummy3[:]
+        res4 = dummy4[:]
+
+    with netCDF4.Dataset(tmp_local_netcdf, "r") as ds:
+        # don't mask, since h5netcdf doesn't do masking
+        if netcdf_write_module == netCDF4:
+            ds.set_auto_mask(False)
+
+        f = ds["test"]
+
+        np.testing.assert_allclose(f.variables["dummy1"][:], res1)
+        assert f.variables["dummy1"].shape == (3, 3)
+        np.testing.assert_allclose(f.variables["dummy2"][:], res2)
+        assert f.variables["dummy2"].shape == (3, 3)
+        np.testing.assert_allclose(f.variables["dummy3"][:], res3)
+        assert f.variables["dummy3"].shape == (3, 3)
+        np.testing.assert_allclose(f.variables["dummy4"][:], res4)
+        assert f.variables["dummy4"].shape == (3, 3)
+
+    with legacyapi.Dataset(tmp_local_netcdf, "r") as ds:
+        f = ds["test"]
+        np.testing.assert_allclose(f.variables["dummy1"][:], res1)
+        assert f.variables["dummy1"].shape == (3, 3)
+        assert f.variables["dummy1"]._h5ds.shape == (3, 3)
+        np.testing.assert_allclose(f.variables["dummy2"][:], res2)
+        assert f.variables["dummy2"].shape == (3, 3)
+        assert f.variables["dummy2"]._h5ds.shape == (1, 3)
+        np.testing.assert_allclose(f.variables["dummy3"][:], res3)
+        assert f.variables["dummy3"].shape == (3, 3)
+        assert f.variables["dummy3"]._h5ds.shape == (2, 3)
+        np.testing.assert_allclose(f.variables["dummy4"][:], res4)
+        assert f.variables["dummy4"].shape == (3, 3)
+        assert f.variables["dummy4"]._h5ds.shape == (0, 3)
+
+    with h5netcdf.File(tmp_local_netcdf, "r") as ds:
+        f = ds["test"]
+        np.testing.assert_allclose(f.variables["dummy1"][:], res1)
+        assert f.variables["dummy1"].shape == (3, 3)
+        assert f.variables["dummy1"]._h5ds.shape == (3, 3)
+        np.testing.assert_allclose(f.variables["dummy2"][:], res2)
+        assert f.variables["dummy2"].shape == (3, 3)
+        assert f.variables["dummy2"]._h5ds.shape == (1, 3)
+        np.testing.assert_allclose(f.variables["dummy3"][:], res3)
+        assert f.variables["dummy3"].shape == (3, 3)
+        assert f.variables["dummy3"]._h5ds.shape == (2, 3)
+        np.testing.assert_allclose(f.variables["dummy4"][:], res4)
+        assert f.variables["dummy4"].shape == (3, 3)
+        assert f.variables["dummy4"]._h5ds.shape == (0, 3)
 
 
 def test_creation_with_h5netcdf_edit_with_netcdf4(tmp_local_netcdf):